Pragmatic metrics (#10)
Instead of taking over responsibility for the `page-visited` metric, we send a new `loaded` metric. While this is confusing and will most likely be asymmetrical when mobile ships, the benefit is an easier transition to the new architecture, without heavy lifting by the metrics processing team.

We also now maintain compatibility with the `browserUserAgent` and `clientTimeMs` field names. The `metricsVersion` field on the other hand, is still enhanced with semver.